requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-image>=0.18.0
geopandas>=0.9.0
shapely>=1.7.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
# OR TensorFlow (choose one)
# tensorflow>=2.6.0

# Image Processing
opencv-python>=4.5.0
opencv-python-headless>=4.5.0
rasterio>=1.2.0
pillow>=8.3.0

# Data Collection
earthengine-api>=0.1.290
requests>=2.25.0

# Web Framework
Django>=4.0.0
djangorestframework>=3.12.0
# Flask for plant disease API
flask>=2.0.0
flask-cors>=3.0.0

# Machine Learning
xgboost>=1.5.0


# Utilities
python-dotenv>=0.19.0
joblib>=1.1.0